# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CRM monitoring API for arbitrage trading system.",
        "title": "Arbitrage VM CRM API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/equity/latest": {
            "get": {
                "description": "Temporarily fetch latest wallet/equity directly from exchange APIs with a short backend cache",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Equity"
                ],
                "summary": "Latest equity from exchange API",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by exchange: bybit, bitget. Supports comma-separated values.",
                        "name": "exchange",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "USDT",
                        "description": "Margin quote coins. Supports comma-separated values.",
                        "name": "quote",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Bypass backend cache and pull from exchange API",
                        "name": "refresh",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.EquityLatest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/response.EquityLatest"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.CombinedEquity": {
            "type": "object",
            "properties": {
                "account_equity": {
                    "type": "string"
                },
                "available_balance": {
                    "type": "string"
                },
                "initial_margin": {
                    "type": "string"
                },
                "maintenance_margin": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "unrealized_pnl": {
                    "type": "string"
                },
                "wallet_balance": {
                    "type": "string"
                }
            }
        },
        "response.EquityCoin": {
            "type": "object",
            "properties": {
                "available_balance": {
                    "type": "string"
                },
                "coin": {
                    "type": "string"
                },
                "equity": {
                    "type": "string"
                },
                "locked": {
                    "type": "string"
                },
                "unrealized_pnl": {
                    "type": "string"
                },
                "usd_value": {
                    "type": "string"
                },
                "wallet_balance": {
                    "type": "string"
                }
            }
        },
        "response.EquityFetchError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "exchange": {
                    "type": "string"
                }
            }
        },
        "response.EquityLatest": {
            "type": "object",
            "properties": {
                "cache_ttl_seconds": {
                    "type": "integer"
                },
                "combined": {
                    "$ref": "#/definitions/response.CombinedEquity"
                },
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.EquitySnapshot"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.EquityFetchError"
                    }
                }
            }
        },
        "response.EquitySnapshot": {
            "type": "object",
            "properties": {
                "account_equity": {
                    "type": "string"
                },
                "available_balance": {
                    "type": "string"
                },
                "cached": {
                    "type": "boolean"
                },
                "coins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.EquityCoin"
                    }
                },
                "exchange": {
                    "type": "string"
                },
                "initial_margin": {
                    "type": "string"
                },
                "maintenance_margin": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "unrealized_pnl": {
                    "type": "string"
                },
                "wallet_balance": {
                    "type": "string"
                }
            }
        },
        "response.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    }
}